refactor: 참가자 관련 패키지 분리
