+++ backend/src/main/java/com/pickeat/backend/participant/application/ParticipantServiceV2.java
@@ -1,12 +1,12 @@
-package com.pickeat.backend.pickeat.application;
+package com.pickeat.backend.participant.application;
 
 import com.pickeat.backend.global.exception.BusinessException;
 import com.pickeat.backend.global.exception.ErrorCode;
 import com.pickeat.backend.login.application.dto.response.TokenResponse;
-import com.pickeat.backend.pickeat.application.dto.request.ParticipantRequestV2;
-import com.pickeat.backend.pickeat.domain.ParticipantV2;
+import com.pickeat.backend.participant.application.dto.request.ParticipantRequestV2;
+import com.pickeat.backend.participant.domain.ParticipantV2;
 import com.pickeat.backend.pickeat.domain.PickeatV2;
-import com.pickeat.backend.pickeat.domain.store.ParticipantStorage;
+import com.pickeat.backend.participant.domain.storage.ParticipantStorage;
 import com.pickeat.backend.pickeat.domain.store.PickeatStorage;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
