+++ backend/src/main/java/com/pickeat/backend/participant/application/ParticipantTokenProvider.java
@@ -1,9 +1,9 @@
-package com.pickeat.backend.pickeat.application;
+package com.pickeat.backend.participant.application;
 
 import com.pickeat.backend.global.auth.JwtProvider;
 import com.pickeat.backend.login.application.dto.response.TokenResponse;
-import com.pickeat.backend.pickeat.domain.ParticipantV2;
-import com.pickeat.backend.pickeat.domain.PickeatV2;
+import com.pickeat.backend.participant.domain.Participant;
+import com.pickeat.backend.pickeat.domain.Pickeat;
 import io.jsonwebtoken.Claims;
 import java.util.HashMap;
 import java.util.Map;
@@ -11,30 +11,28 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
 @Component
-public class ParticipantTokenProviderV2 {
+public class ParticipantTokenProvider {
 
     private static final String PICKEAT_CODE_CLAIM_KEY = "pickeatCode";
 
     private final JwtProvider jwtProvider;
     private final long expirationMillis;
 
-    public ParticipantTokenProviderV2(
-            JwtProvider jwtProvider,
-            @Value("${participant.jwt.expiration}") long expirationMillis
-    ) {
+    public ParticipantTokenProvider(JwtProvider jwtProvider,
+            @Value("${participant.jwt.expiration}") long expirationMillis) {
         this.jwtProvider = jwtProvider;
         this.expirationMillis = expirationMillis;
     }
 
-    public TokenResponse createToken(ParticipantV2 participant, PickeatV2 pickeat) {
+    public TokenResponse createToken(Participant participant, Pickeat pickeat) {
         Map<String, Object> claims = new HashMap<>();
-        claims.put(PICKEAT_CODE_CLAIM_KEY, pickeat.getCode());
-        return jwtProvider.createTokenWithClaims(participant.getCode(), expirationMillis, claims);
+        claims.put(PICKEAT_CODE_CLAIM_KEY, pickeat.getCode().getValue().toString());
+        return jwtProvider.createTokenWithClaims(participant.getId(), expirationMillis, claims);
     }
 
-    public String getParticipantCode(String token) {
+    public Long getParticipantId(String token) {
         Claims claims = jwtProvider.getClaims(token);
-        return claims.getSubject();
+        return Long.parseLong(claims.getSubject());
     }
 
     public String getPickeatCode(String token) {
