+++ backend/src/main/java/com/pickeat/backend/participant/domain/ParticipantV2.java
@@ -1,4 +1,4 @@
-package com.pickeat.backend.pickeat.domain;
+package com.pickeat.backend.participant.domain;
 
 import java.util.UUID;
 import lombok.AccessLevel;
