+++ backend/src/main/java/com/pickeat/backend/participant/domain/repository/ParticipantRepository.java
@@ -1,10 +1,11 @@
-package com.pickeat.backend.pickeat.domain.repository;
+package com.pickeat.backend.participant.domain.repository;
 
-import com.pickeat.backend.pickeat.domain.Participant;
+import com.pickeat.backend.participant.domain.Participant;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ParticipantRepository extends JpaRepository<Participant, Long> {
+
     List<Participant> findByPickeatIdIn(List<Long> pickeatIds);
 
     List<Participant> findByPickeatId(Long pickeatId);
