+++ backend/src/test/java/com/pickeat/backend/participant/application/ParticipantServiceTest.java
@@ -1,4 +1,4 @@
-package com.pickeat.backend.pickeat.application;
+package com.pickeat.backend.participant.application;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
@@ -6,10 +6,10 @@ import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import com.pickeat.backend.global.auth.JwtProvider;
 import com.pickeat.backend.global.exception.BusinessException;
 import com.pickeat.backend.global.exception.ErrorCode;
-import com.pickeat.backend.pickeat.application.dto.request.ParticipantRequest;
-import com.pickeat.backend.pickeat.domain.Participant;
+import com.pickeat.backend.participant.application.dto.request.ParticipantRequest;
+import com.pickeat.backend.participant.domain.Participant;
 import com.pickeat.backend.pickeat.domain.Pickeat;
-import com.pickeat.backend.pickeat.domain.repository.ParticipantRepository;
+import com.pickeat.backend.participant.domain.repository.ParticipantRepository;
 import com.pickeat.backend.support.DatabaseSliceTest;
 import com.pickeat.backend.support.fixture.ParticipantFixture;
 import com.pickeat.backend.support.fixture.PickeatFixture;
