+++ backend/src/test/java/com/pickeat/backend/participant/domain/ParticipantTest.java
@@ -1,7 +1,8 @@
-package com.pickeat.backend.pickeat.domain;
+package com.pickeat.backend.participant.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import com.pickeat.backend.pickeat.domain.Pickeat;
 import com.pickeat.backend.support.fixture.PickeatFixture;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
